base: Compare Events based on the whole payload
Due to some, to me yet unknown, way the comparison yields incorrect
results while creating new matches in BinaryNode. The new potential
chain of matches are visibly distinct when their whole payload is
compared but comparing the counter shows that they are, in fact, equal.
This is just wrong and as an immediate fix, compare the whole payload
instead of just the counter (i.e., ID).