misc: Add basic package for tracking metrics
This is a very basic module that adds simple methods to track some
metrics. Most of the work here will go into properly visualizing the
results.

Initially I thought of using OpenTelemetry [0] but the Python SDK is
very cloud-service centric and is not a good fit for your general
small-scale local application.

[0] https://opentelemetry.io